ref: move command protocols to namespace
