+++ src/CoreLib/Game/Commands/Protocols/CommandPunishmentProtocol.cs
@@ -4,7 +4,7 @@ using Imlight.CoreLib.WizardData.Models;
 using System;
 using System.Text;
 
-namespace Imlight.CoreLib.Game.Commands;
+namespace Imlight.CoreLib.Game.Commands.Protocols;
 
 internal class CommandPunishmentProtocol : CommandProtocol {
     internal override string Group { get; set; } = "";
